LMS-1915 microscopy data upload, LMS-1920 depth-scans upload
SVN: 19059

+++ screening/sourceTest/java/ch/systemsx/cisd/openbis/plugin/screening/shared/imaging/FeatureTableBuilderTest.java
@@ -139,7 +139,7 @@ public class FeatureTableBuilderTest extends AssertJUnit
                 {
                     String dataSetCode = "ds" + dataSetID;
                     one(dao).tryGetDatasetByPermId(dataSetCode);
-                    int containerId = dataSetID + 100;
+                    long containerId = dataSetID + 100;
                     ImgDatasetDTO dataSet =
                             new ImgDatasetDTO(dataSetCode, null, null, containerId, false);
                     dataSet.setId(dataSetID);
